services/worker/requirements.txt
qdrant-client==1.9.1
langchain==0.3.3
langchain-community==0.3.2
python-docx==1.1.2
pypdf==4.3.1
python-dotenv==1.0.1